# src/main/java/com/nelumbo/parqueadero_api/controller/UserController.java
package com.nelumbo.parqueadero_api.controller;

import com.nelumbo.parqueadero_api.dto.AdminVehicleResponseDTO;
import com.nelumbo.parqueadero_api.dto.ParkingResponseDTO;
import com.nelumbo.parqueadero_api.dto.UserRequestDTO;
import com.nelumbo.parqueadero_api.dto.UserResponseDTO;
import com.nelumbo.parqueadero_api.services.ParkingService;
import com.nelumbo.parqueadero_api.services.UserService;
import io.swagger.v3.oas.annotations.Operation;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/users")
@RequiredArgsConstructor
public class UserController {

    private final UserService userService;
    private final ParkingService parkingService;

    @PostMapping
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<UserResponseDTO> createUser(@Valid @RequestBody UserRequestDTO userRequest) {
        UserResponseDTO createdUser = userService.createUser(userRequest);
        return new ResponseEntity<>(createdUser, HttpStatus.CREATED);
    }



    @GetMapping("/socio/parkings")
    public List<ParkingResponseDTO> getParkingsBySocio(
            @AuthenticationPrincipal UserDetails userDetails) {

        return parkingService.getParkingsBySocio(userDetails.getUsername());
    }




//de cualquiera rol admin
    @GetMapping("/parkings/{parkingId}/vehicles")
    public List<AdminVehicleResponseDTO> getVehiclesInParking(
            @PathVariable Integer parkingId,
            @RequestParam(required = false) Boolean activeOnly) {

        return parkingService.getVehiclesInParking(parkingId);
    }

    }